fix EncapsulateFieldTests
Factory_NullSelectionReturnsNullPresenter relied on the implementation
for GetQualifiedSelection() of the actual code. Mocks didn't follow suit
and returned non-null QualifiedSelection for empty Selection instances.

For whatever reason we can't set the mock up to CallBase() so we need to
special case it. Using a ternary inside the lambda results in a compiler
error