Add status bar texts for project, component, form and control selections
Currently working off RawInput, so some of the existing timings are
already off, but also apply to project and component selections in this
commit. Clicking a Project Explorer component a second time forces the
correct text. This will be fixed once the events are triggered by the
VBE and not RawInput.

The project/component/form/control selection is currently ignorant of
Declarations, so it doesn't know about any references for the selected
item.

If you have a form designer AND its code module open, and the form
component is selected in Project Explorer, but a non-form window has
focus, Rubberduck can't determine which window has the focus, and
assumes it is the codepane. This is the opposite behavior of the VBA
Properties window, but RD has more to offer in code modules, so for now,
this is a known issue.

And finally, it's not hooked up to the Rubberduck Code Explorer, yet.