Adds check to if a userform selected in Project Explorer has a designer that is topmost in MDI
If the user selects various components in Project Explorer, and then
selects a userform, that component might already have an open designer
that is visible and top-most in the MDIClient. If the designer is
topmost, then set the status bar to the designer's control selection,
otherwise use the component.
If the userforms's code module is topmost, it has already been hit
before we reach this check