First step towards clean exit (#3277)
This PR is a first step towards a clean exist; it fixes one reason of the crash on exit: we have not been unregistering the Click events from the CommandBarButtons, neither in the command bar nor in the menus. This leads to a fatal error on shutdown.

This PR changes the CommandBarControls to be added permanently and to be removed explicitly on shutdown, removing the Click handlers as well. Moreover, as a safety measure, the GC gets invoked at the end of the shutdown.

There are more problematic sections I have identified and I am looking into. However, although there are still plenty of COM access violations on shutdown, I get a clean exit in Excel in production now, but not when debugging.

ref. https://github.com/rubberduck-vba/Rubberduck/issues/3213

* Make menus and the command bar permanent and clean them up on shutdown.

* Fix event handler registration in CommandBarButton.

* Fixed RemoveChildren on the parent menus and the command bar.

* Add garbage collection on shutdown.

* Rewired deleting the command bar and menus.

* Fixed three tests with illegal code as input.