Add InternalsVisibleTo for RubberduckTests to all other projects,
We want to be able to restrict type visibility as much as possible for all projects,
especially so if we want to execute arbitrary user-code (plugins).
For that purpose as many types as possible should be marked internal.

To still be able to test these types, they must be visible to the test-project.
See #1600