Make TestEngine more responsible
LastRun and running by recorded outcome were cluttering the TestExplorerModel.
Both these responsibilities have been transferred to the TestEngine as the
single source of truth for the outcomes and run state of tests.

This implies that the DelegateCommands and methods used in the TestExplorerViewModel
are now full-fledged classes of their own. They are accordingly extracted into their own
classes in a separated namespace to avoid cluttering the Commands namespace