Enable nested scrolling while in settings panes
Scrolling in the settings windows will now bubble up to the next parent. When a child reaches the top/bottom extent of its scrolling area it raises the event on the parent element to continue scrolling.

This has only been tested with a single nested ScrollViewer. Further nesting levels have not been verified to work.