Make the state after parse of no components the state Ready.
In case there are no components, leaving the state at Pending parsing provides the impression to the user that nothing has processed and that RD does not work. After testing, nothing really breaks in the Ready state that does not have the same broken behaviour when returning to Pending after a parse.