Create new project declarations before parsing and refresh project references
Previously, we coupled the generation of project declarations to that of its components. That causes problems in case of projects without components.

Moreover, we replaced the project declarations with new ones on each reparse of one of its components, even in case not all components were replaced, leaving the unchanged component references with the old project declaration as parent.

Finally, we did not refresh the project references on the project declarations in case of changes to the project references in the VBE. This had no impact on which references were loaded but the references got stale on not recreated project declarations.