Introduce more annotation classes and more fine-grained legality
In the previous setup, only member and module annotations existed. That is a problem for annotations (also) applicable to other targets, like variables or identifier references.

This commit has two failing tests that uncover that annotations on module level variables are not treated correctly: they get added to the module as well.