Expose the IRewritingManager on the RubberduckParserState
This change is to be reverted later. It allows to decouple the work necessary to change the users of the IModuleRewriter to use the IRewritingManager from the work necessary to constructor inject the IRewritingManager into them. (The latter is a lot of work in the tests.)

Here, property injection is used because the construction of the RewritingManager ultimately depends on the RubberduckParserState to be there.