Add safety around compilation argument changes between parses
Previously, a further parse request at the wrong time would have caused the information to be lost that modules have to be reparsed due to a change of compilation arguments. To avoid this, these get saved in a hash set on the ParseCoordinator that gets cleared after a complete parse.