Exract loading declarations from ComProject
Previously, loading declarations from ComProjects was handled by a method in the abstract base class ReferencedDeclarationsLoader. This is a problem because it couples loading the declarations from a ComProject to loading them from a reference.

Now, that this is extracted into its own strategy it can be used for user ComProjects as well.