Change how attribute annotations work
Now they have different properties for the attribute and the values. This makes validating whether the attributes are in sync with the annotations considerably easier.

Moreover, whether a module attribute annotations is deemed to have a matching attribute no longer depends on the value of the attribute. That was a problem for developing an AddMissingAttributeQuickfix as it would duplicate attributes otherwise.