Make attribute annotations use the attributes on the annotation types
Previously, there was no connection between the attributes on the annotation types and the actual attributes. Now, the attribute annotations determine how to get their values based on the attributes.

There are three different types of IAttributeAnnotations: the FlexibleAttributeAnnotationBase subclasses using the first parameter as attribute name and the remaining as values, the FixedAttributeValueAnnotationBase subclasses ignoring the values and taking both the attribute name and the values from the attribute on the annotation type, and the FlexibleAttributeValueAnnotation in between taking the attribute name and the number of values from the attribute on the annotation type, but the actual values from the annotation values.