Add the MemberAttributesRecoverer
On this class, one can request the attributes of members to be recovered on the next parse. It will do this after the declarations have been resolved. In order to not process information with the attributes still missing, it cancels the current parse right after requesting its own attributes rewrite.