Invalidate rewrite sessions on checkout of a dirty rewriter
A rewriter can only be dirty on checkout if the code of the corresponding module has been modified since the last parse. Applying such a rewriter will erase all modifications since the last parse and thus probably destroy work by the user. To avoid this, the rewrite sessions will invalidate themselves when they encounter a stale module.

To allow to communicate the reason of the invalidation, IsInvalidated and Invalidate have been replaced by a Status property taking values in the new RewriteSessionState enum. It can only be changed from Valid to some other state, Afterwards, all attempts to set it are ignored. Consequently, the Status shows the original reason for invalidation.