Add a mechanism to ask an inspection result whether it gets invalidated
This commit introduces the method ChangesInvalidateResult on the IInspectionResult that returns whether a change of the specified modules makes the result invalid. The base implementation returns true if the module specified on the result gets invalidated or else forwards the request to the inspection to allow inspection specific handling.

Declaration- and IdentifierReferenceInspectionResults also get invalidated based on the state they carry. AggregateInspectionResults are always considered invalidated as there is no way to see the individual results.