Add code pane open state recovery to the ISelectionRecoverer
This allows to save that code panes are open and to reopening them. Reopening has the side-effect of activating the modules. So, care has to be taken when also recovering the active code pane. This has to happen after reopening the code panes.

To avoid head aches around the sequencing of event handlers, reopening code panes on the next parse is not supported.