Push the IRewritingManager and ISelectionService to RefactorCommandBase
All refactoring commands but the not yet remodeled ExtractMethod command use these two, which are essential for the refactorings. So, I have pushed them to the base class.

This also removes the IVBE from the base class since it is no longer needed by the majority of the refactoring commands.