Introduce RefactoringTargetExceptions
These exceptions are used by the different refactorings to signal failures (based on the input) to the caller. There are general ones like NoActiveSelectionException and specialised ones for particular refactorings, e.g. the TargetDeclarationIsAlreadyAFieldException for the IntroduceFieldRefactoring.

This also fixes a bug in IntroduceFieldRefactoring in the presence of more than one module body element.