Refactor the RenameRefactoring
The model now stores whether and why a derived target had to be chosen. This allows to push querying for confirmation to the presenter, which does this before attempting to show the main dialog.

Moreover, the view model now evaluates on click of the OK button whether there is a name conflict and asks for confirmation to proceed nontheless.

These two changes allow to remove all UI from the refactoring itself.

In addition, the logic to get the target declaration from a selection has been pulled from the model to the refactoring. This allows to have to expected call order of the overloads of Refactor in the refactoring and makes the model a pure data storage.