Enforce base class CanExecute in derived Command classes
Also removes no longer needed base constructor calls passing the current class logger to CommandBase.

Also fixes a bug in the CanEvaluate method of the IndentCommand.

This breaks the AddTemplateCommand since it does not adhere to the can evaluate contract of its base class. (Needs to be separated in a further commit.)