Remove Inspection Result Aggregation
The original rationale for inspection result aggregation was the excessive memory use
when displaying them. Since the rework of the Inspections view, this is no longer an issue.

As such aggregation, which has multiple issues is removed in this commit. This change makes
the following issues completely obsolete and fixes them that way:

fixes #4561, fixes #3597, fixes #3172, fixes #2900, fixes #4870