Inverted dependency graph for Rubberduck.CodeAnalysis,
For that purpose we cleaned up Settings and their loading behaviour a bit and moved most of the interfaces relating to it
into Rubberduck.SettingsProvider. The Defaults are now loaded differently, registration is probably utterly broken, though.

Furthermore this extracts the CodeInspectionSettings defaults to RD.CodeAnalysis to make them accessible for the loader there.