Make SettingsProvider responsible for settings
This moves the Consumer-Visible interfaces for Settings into the SettingsProvider project.
Additionally a base-implementation for a ConfigurationService is provided.
It uses a clear reflection-based mechanism to load the defaults for a given settings type
from a settings class. The type is provided as a Type-Argument to the constructor.

This structure should allow easily extensible and customizable settings loading with caching and defaults.