Create SearchBox component
This component replaces the search boxes in CodeExplorer and Add/Remove References dialog.
It provides a vast simplification over the preexisting use of overriding a TextBox with a
Style defined in the resources section.

Also this commit moves Behaviours into their own solution folder, though no namespace adjustments
are performed.