Implemented this inspection as DeclarationInspectionResult, instead of QualifiedContextInspectionResult.
Got rid of all refrences to redundant ImethodContext interface from previous commit.

Added a little extension method to return DeclarationType.PropertyGet formatted as "Property Get" instead of "Propertyget". If there is a better way I would love to hear.
In this context alone it seems a bit redundant, but maybe if we move it to Rubberduck.Parsing.Symbols it may serve others as well.

Finally, added an abstract body context to ModuleBodyElementDeclaration, because all these declarations should have a body, and it made inspection easier.