Prefer builtInType over lExpr in expression in grammar
Previously, everything matched by builtInTypeExpr was already consumed by lExpr. Since all built-in types are keywords, this should not cause problems in most cases. The only exception would be foreign identifiers whose name is a built-in type. (This should be very rare and is problematic anyway.)

This change required patching up the resolver as well to deal with builtInTypes.