Fix formatting in RD command bar broken by prior commits
This includes a complete refactoring of ContextFormatter and changes the behaviour to always show parentheses in in the type if the declaration is an array. Moreover, the parentheses following the identifier name are no longer added for array variables. (They are obviously arrays based on the new parentheses in the type name.)