Redesign Annotation Processing
This redesigns annotation processing to no longer rely on the AnnotationTypes enum to identify
an annotation, allowing us to separate the scoping of an annotation from it's type.

Additionally this explicitly stores metainformation on annotations in Attributes on the specific
annotation type. This metainformation is used to automatically register annotations for the
code pane parsing process.

These changes have far-reaching implications for how annotations are used, most of them addressed
in this commit. The parsing of annotations through the VBA API is not correctly dealt with.