Second redesign for annotations
This moves static information on annotations into implementations of IAnnotation
and introduces a ParseTreeAnnotation class to correlate parser contexts to annotations.
With that change we get a clearer separation between the annotation and it's use.

For workability, a handful of extension methods have been added, some have been adapted.
Additionally a large number of predicates and filterings have been adjusted to the new API.