Move refactorings away from FindVariable
Instead, an ISelectedDeclarationProvider is used to find the variable to refactor. This has the effect that the refactorings can no longer be executed for some selections, e.g. for selections containing the As type clause. This is in line with the declaration shown in the status bar.

This affects IntroduceField, IntroduceParameter, EncapsulateField and MoveCloserToUsage.