Move RegexSearchAndReplace awai from FindTarget and make it a bit more sane
This fixes the method to get the surrounding block in SearchInCurrentBlock and generally fixes how searching within a given selection works. Instead of filtering the results for the entire module, the search is performed only inside the selection. This makes a difference on the start and end line.

Moreover, this commits adds fixmes for the broken handling of code modules, whose root cause is saving the modules on the results. This should never be done.