Fix deadlock when coming out of suspension on the UI thread in Ready state
On the change to Ready state, the CE dispatches work to the UI thread and waits for the completion. If the change was triggered from the UI thread, this causes a deadlock. Usually, such a change only happens during parsing, which is executed on a background thread. However, when coming out of parser suspension, it can happen as well. To deal with this, the state changed action now also checks that the old state was not Busy.