Make TestEngine no longer available outside the Ready state to prevent deadlocks
The TestEngine suspends the parser from the UI thread, which deadlocks on the CE update on the change to the Ready state. Since we cannot execute anyway before the parse has completed due to the suspension attempt disallowing to trigger the tests outside the Ready state is not really a great step back.