Move checks for import into ImportCommand
Previously, these checks were only executed in the update from files command. However, it was discovered that not doing checks will silently not import modules for which the import fails and even aborts in the middle for document modules.

Because of these problems, all major checks are now executed for all subtypes of the import command. The subtypes now only determine the modules to remove before import and are allowed to add some more validation at chosen places in the import checks.