Update extract public interface
A private class can have a (public | private) interface extracted. The
default option is public.

When an implementing class is public the newly created interface must also
too be public. As such the private option for the extracted interface is
disabled.
Need to get the Public.Checked = true on private implementing classes. Once
that's done I can merge the following commits together and edit this commit.