Add more high level exception handling
The async void responsible for refreshing the inspection results now has a to level exception handler swallowing all exceptions. This is necessary since bubbling out of there will likely take down the host.

Moreover, we now log and swallow COM exceptions coming from the process of changing the status label of the refresh button. This means that COM registration issues for the command bar controls will no abort  the calling process. The only effect will be that the label does not change.