Introduce high priority state changed event
The order in which event handlers are executed is not guaranteed. Because of this we had the problem that the status message change for the state change to Ready happened after the status message change in the inspections, which are executed async. As the result, the message 'Inspection' was immediately overwritten and did not last until the inspections had  finished.