Replace the dynamic PropertyBag on inspection results
It is replaced by a generic method and a dedicated property for the names of disabled quick fixes. The three kinds of inspection results now also have one generic subtype each that takes an additional properties value of the generic type and returns it via the corresponding generic properties method.

The consumer of the result properties still has to know the type of properties and what they represent (We want to avoid circular dependencies to the quick fixes.), but the properties can now be provided in a strongly typed fashion and lend themselves better to generation in inspection base classes.