Redesign ParameterNotUsedInspection
For interfaces, it now only reports parameters of the interface member and not of individual implementations. Moreover, it does it only if there are implementations and none uses the parameter.

The same behaviour now also applies to user events and event handlers. (Parameters ofevent handlers of built-in events are never reported.)

In addition, ClassModuleDeclaration.IsInterface has been enhanced to also retun true is there is an @Interface annotation on the class.