Add base classes for declaration inspections that use global information
The information gets passed down to the method determining whether a declaration is a result. This means that the signatures of the methods are different from the the other base classes. So, these for a second hierarchy.

This also changes the base classes to also look at the QMNs of user projects.

Finally, this makes the final declaration inspections use the base classes.