Introduce CodeExplorerMoveToFolderCommand
This combines MoveToFolder and MoveFolder, selecting based on the selected node of the CE.
In principle, the backing refactoring actions are capable of dealing with multiple nodes at once. However, the CE only allows selecting single items at the moment and the command reflects that when deriving the models to pass to the refactoring actions.