Fix WithMemberAccesses in MemberMayReturnNothingInspection
This caused an NRE.

In addition, a warning (doc-)comment is added to indicate that this base class is not suitable for members that can be accessed unqualified; there simply is no branch handling this case and it would either not work or even find the wrong usage context.

This restriction rules out all user code and all members on global instances as well as on document and form classes.