Add delay to search box input
This adds a delay of 400ms to the search boxes in the test explorer, the inspection results window and the CE. This means that a somewhat below average typer with a typing speed of 150cpm will just not trigger a UI refresh while typing.

In particular in the CE, the delay is considerably shorter than the refresh time for the filter change.