Fix fix menu in inspection results window
The relative source, which works for the context menu, does not for the toolbar menu because the grouping grid is in another branch of the logical tree. Instead, the grouping grid can be referred by name, which does not work for the grouping grid's context menu. Accordingly, a solution with a fallback is employed.