Get document supertypes from user com projects
We use the user com projects to get the names of the implemented interfaces; the type hierarchy pass does the rest.

There is one hack in this: we suppress the underscore in front of the implemented interfaces in order to emulate the VBE's behaviour to let type checks pass relative to the coclass instead of the interface. This is entirely based on convention.

Moreover, this lacks unit test since the com projects cannot be mocked at this stage.