Add ImplicitContainingWorkbookReferenceInspection and ImplicitContainingWorksheetReferenceInspection
These two find unqualified references inside the corresponding document modules to certain members that point to the ActiveWorkbook or ActiveSheet when used unqualified outside the appropriate document modules. That the reference is actually to the containing document can be surprising. Because of that, they should be qualified with Me.

This also removes the corresponding false-positives for the ImplicitActiveSheetReferenceInspection and ImplicitActiveWorkbookReferenceInspection.

Technically, both the containing and active workbook/worksheet types now share a common base inspection handling (most of) the selection of the declarations whose references can be suspicious.