Fix missing default member result for optional parentheses on properties
The problem was that the empty argument list gets parsed as an argument list with a single missing argument and the IndexDefaultBinding did not account for it. Since the missing argument is actually a valid way to parse the empty argument list and the parser cannot know whether there is an optional parameter or not, I chose to catch this in the IndexDefaultBinding.