Add IgnoreInModule annotation in result module and not target module
The target of an inspection result is the declaration it refers to. For identifier reference results, this is not necessarily located in the same module as the identifier reference. To always get the correct module, I have switched it to the module on the result's qualified selection. This should exist for all results referring to some location in code.