Add type resolution for ReDim declared arrays
Previously, any as type clause or type hint was ignored, leading to inspection false positives and wrong resolution of members of array elements.

The solution adds a copy of the type resolution on the spot where the ReDim declared arrays are added as pseudo-undeclared variables. This requires to generate the expression parser and binding service each time and to hardcode them. This is not ideal, but it should be solved when extracting the entire logic for undeclared variables out of the declaration finder.