Fix FunctionReturnValueDiscardedInspection for member access on parameterless call
The problem not correctly handled before was that the reference in a member access expression goes to the unrestricted identifier while it goes to the whole simple name expression in case of a simple name expression. This makes it necessary to deal with them in different ways; the first member access parent of a reference to a member access is the member access itself while it is a further member access for all other kinds of references.