Correctly resolve references to declarations named like built-in types
The problem is that expressions prefer to resolve built-in type names as built-in types in favor of simple name expressions in order to be able to distinguish them. Unfortunately, VBA allows to name declarations like built-in types.

To fix this, whenever the default expression binding context encounters a built-in type, it first tries to bind it as a simple name. It tries to resolve it and if that is successful, it returns the binding. Otherwise, it falls back to the built-in type binding.