Add non-disposal decorators for com wrappers
This adds decorators for several com wrapper types that suppress disposal attempts on them.
This is relevant in two places. First, the wrappers handed out by the ProjectsRepository are references to cashed wrappers; only the repository itself should be allowed to dispose the instances when it clears its cache. Second, disposal of the com wrappers bound in CW is non-deterministic. In order to avoid running into concurrency issues on shutdown, CW should not be allowed to dispose the wrappers. This will than happen deterministically when disposing the com safe.