Make UnassignedVariableUsageInspection aware of redimmed variant arrays
The problem was that an array declares ad a Variant and later assigned vie ReDim is not flagged as an array and thus not properly excluded from the inspection.

The approach taken here is to explicitly look for ReDim before the usage inside the same function as the reference. This approach has been taken in favor of enhancing the resolver because whether a Variant declaration is an array depends on the exact location in the code. One can define a Variant then redim it as an array and later assign a non-array value to it. This makes it comparatively complicated to figure out whether a Variant is an array in every spot. Anyway, we store the information whether a variable is an array on the declaration and this does not allow changing values over the code.

Note that that this fix goes in a conservative direction of removing false positives in favor o introducing more obscure false negatives. In particular, the case is no longer covered where the ReDim is conditional, i.e. does not happen in all branches of the code execution.