Align which unused procedures are reported
Now, implemented interface members and members in class modules marked as interface are reported as unused, if there are no references to them.

In addition, we no longer report unused procedures in in exposed class modules, sine these modules are clearly intended for external use.
We do not treat standard modules without the Private Module option the same way, because not having the option is the default in contrast to the exposed attribute.