Handle inconsistent results from Find All references gracefullt
The results of Find All References take the input code from the reference's line in the VBE and the highlight selection based on the reference selection. If this happens in a dirty state, the highlight can be outside the input code on the line that is now on the line the reference was on before. Previously, this caused an ArgumentOutOfBoundExcetion in the UI code when applying the highlight, which brought down the host. Now, we check whether the highlight is within the input code and only apply it if this check passes.

The navigation properly navigates to the last known selection of the reference.