Try to avoid RCW exception at shutdown
This wraps the calls to set and get the enabled state, the visibility and the tag of a CommanBarButton in a check whether it has already been released an logs any attempt the access these methods after the release as a warning.

This should lead to a clean exit. However, it only fixes the symptoms of apparently too late CanExecute checks. Identifying these is made difficult by the fact that setting the enabled state is dispatched to the UI thread, which cuts off the stack trace.