Replace Exit Sub when converting Subs to Functions
This assigns the return value and then exists the function instead of exiting the procedure that gets converted. These two statements are inserted separated by a statement separator instead of doing it on multiple lines to avoid having a separate code path if the Exit Sub is in a single line if statement.